test: test for pendingOrderActionMiddleware
