+++ src/legacy/state/orders/helpers.tsx
@@ -1,10 +1,8 @@
-import styled from 'styled-components/macro'
-
-import { formatOrderId, shortenOrderId } from 'legacy/utils'
+import { formatOrderId } from 'legacy/utils'
 import { OrderID } from 'api/gnosisProtocol'
 import { addPopup } from 'legacy/state/application/reducer'
 import { OrderStatus } from './actions'
-import { CancellationSummary } from 'modules/account/containers/Transaction/styled'
+import { OrderObject, OrdersStateNetwork } from './reducer'
 
 type OrderStatusExtended = OrderStatus | 'submitted' | 'presigned'
 
@@ -67,28 +65,6 @@ function setOrderSummary({ id, summary, status, descriptor }: SetOrderSummaryPar
   return summary
 }
 
-const Wrapper = styled.div`
-  & > p:first-child {
-    margin-top: 0;
-  }
-
-  & > p:last-child {
-    margin-bottom: 0;
-  }
-`
-
-export function buildCancellationPopupSummary(id: string, summary: string | undefined): JSX.Element {
-  return (
-    <Wrapper>
-      <p>Order successfully cancelled</p>
-      <p>
-        Order <strong>{shortenOrderId(id)}</strong>:
-      </p>
-      <CancellationSummary as="p">{summary}</CancellationSummary>
-    </Wrapper>
-  )
-}
-
 // Metatxn popup
 export function setPopupData(
   type: OrderTxTypes.METATXN,
@@ -132,4 +108,22 @@ export function setPopupData(
   }
 
   return { key, content }
+}
+
+export function getOrderByIdFromState(orders: OrdersStateNetwork | undefined, id: string): OrderObject | undefined {
+  if (!orders) {
+    return
+  }
+
+  const { pending, presignaturePending, fulfilled, expired, cancelled, creating, failed } = orders
+
+  return (
+    pending?.[id] ||
+    presignaturePending?.[id] ||
+    fulfilled?.[id] ||
+    expired?.[id] ||
+    cancelled?.[id] ||
+    creating?.[id] ||
+    failed?.[id]
+  )
 }
